--- entry.js
@@ -1,1 +1,0 @@
-require('./build/index')['default'];
